Fix tests and deprecate PrimitiveValue
