+++ Cargo.toml
@@ -14,7 +14,7 @@ members = [
 
 [package]
 name = "sea-query"
-version = "0.16.3"
+version = "0.17.0"
 authors = [
     "Chris Tsang <tyt2y7@gmail.com>",
     "Billy Chan <ccw.billy.123@gmail.com>",
