+++ src/backend/mysql/table.rs
@@ -96,7 +96,9 @@ impl TableBuilder for MysqlQueryBuilder {
         match column_spec {
             ColumnSpec::Null => write!(sql, "NULL"),
             ColumnSpec::NotNull => write!(sql, "NOT NULL"),
-            ColumnSpec::Default(value) => write!(sql, "DEFAULT {}", self.value_to_string(value)),
+            ColumnSpec::Default(value) => {
+                write!(sql, "DEFAULT {}", value.query_value(&MysqlQueryBuilder))
+            }
             ColumnSpec::AutoIncrement => write!(sql, "AUTO_INCREMENT"),
             ColumnSpec::UniqueKey => write!(sql, "UNIQUE"),
             ColumnSpec::PrimaryKey => write!(sql, "PRIMARY KEY"),
