+++ src/backend/sqlite/table.rs
@@ -116,7 +116,9 @@ impl TableBuilder for SqliteQueryBuilder {
         match column_spec {
             ColumnSpec::Null => write!(sql, "NULL"),
             ColumnSpec::NotNull => write!(sql, "NOT NULL"),
-            ColumnSpec::Default(value) => write!(sql, "DEFAULT {}", self.value_to_string(value)),
+            ColumnSpec::Default(value) => {
+                write!(sql, "DEFAULT {}", value.query_value(&SqliteQueryBuilder))
+            }
             ColumnSpec::AutoIncrement => write!(sql, "AUTOINCREMENT"),
             ColumnSpec::UniqueKey => write!(sql, "UNIQUE"),
             ColumnSpec::PrimaryKey => write!(sql, "PRIMARY KEY"),
