+++ src/expr.rs
@@ -4,7 +4,7 @@
 //!
 //! [`SimpleExpr`] is the expression common among select fields, where clauses and many other places.
 
-use crate::{func::*, query::*, types::*};
+use crate::{func::*, query::*, types::*, value::*};
 
 /// Helper to build a [`SimpleExpr`].
 #[derive(Debug, Clone, Default)]
@@ -28,10 +28,10 @@ pub enum SimpleExpr {
     FunctionCall(Function, Vec<SimpleExpr>),
     Binary(Box<SimpleExpr>, BinOper, Box<SimpleExpr>),
     SubQuery(Box<SelectStatement>),
-    Value(Box<dyn QueryValue>),
-    Values(Vec<Box<dyn QueryValue>>),
+    Value(Value),
+    Values(Vec<Value>),
     Custom(String),
-    CustomWithValues(String, Vec<Box<dyn QueryValue>>),
+    CustomWithValues(String, Vec<Value>),
     Keyword(Keyword),
 }
 
@@ -171,7 +171,7 @@ impl Expr {
     /// ```
     pub fn val<V>(v: V) -> Self
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         Self::new_with_left(SimpleExpr::Value(v.into()))
     }
@@ -237,7 +237,7 @@ impl Expr {
     /// ```
     pub fn value<V>(v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         SimpleExpr::Value(v.into())
     }
@@ -337,7 +337,7 @@ impl Expr {
     /// ```
     pub fn cust_with_values<V, I>(s: &str, v: I) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
         I: IntoIterator<Item = V>,
     {
         SimpleExpr::CustomWithValues(s.to_owned(), v.into_iter().map(|v| v.into()).collect())
@@ -372,7 +372,7 @@ impl Expr {
     /// ```
     pub fn eq<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::Equal, SimpleExpr::Value(v.into()))
     }
@@ -406,7 +406,7 @@ impl Expr {
     /// ```
     pub fn ne<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::NotEqual, SimpleExpr::Value(v.into()))
     }
@@ -477,7 +477,7 @@ impl Expr {
     /// ```
     pub fn gt<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::GreaterThan, SimpleExpr::Value(v.into()))
     }
@@ -510,7 +510,7 @@ impl Expr {
     /// ```
     pub fn gte<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::GreaterThanOrEqual, SimpleExpr::Value(v.into()))
     }
@@ -543,7 +543,7 @@ impl Expr {
     /// ```
     pub fn lt<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::SmallerThan, SimpleExpr::Value(v.into()))
     }
@@ -576,7 +576,7 @@ impl Expr {
     /// ```
     pub fn lte<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::SmallerThanOrEqual, SimpleExpr::Value(v.into()))
     }
@@ -610,7 +610,7 @@ impl Expr {
     #[allow(clippy::should_implement_trait)]
     pub fn add<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::Add, SimpleExpr::Value(v.into()))
     }
@@ -644,7 +644,7 @@ impl Expr {
     #[allow(clippy::should_implement_trait)]
     pub fn sub<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::Sub, SimpleExpr::Value(v.into()))
     }
@@ -678,7 +678,7 @@ impl Expr {
     #[allow(clippy::should_implement_trait)]
     pub fn mul<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::Mul, SimpleExpr::Value(v.into()))
     }
@@ -712,7 +712,7 @@ impl Expr {
     #[allow(clippy::should_implement_trait)]
     pub fn div<V>(self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(BinOper::Div, SimpleExpr::Value(v.into()))
     }
@@ -745,7 +745,7 @@ impl Expr {
     /// ```
     pub fn between<V>(self, a: V, b: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.between_or_not_between(BinOper::Between, a, b)
     }
@@ -778,14 +778,14 @@ impl Expr {
     /// ```
     pub fn not_between<V>(self, a: V, b: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.between_or_not_between(BinOper::NotBetween, a, b)
     }
 
     fn between_or_not_between<V>(self, op: BinOper, a: V, b: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         self.bin_oper(
             op,
@@ -824,11 +824,11 @@ impl Expr {
     /// );
     /// ```
     pub fn like(self, v: &str) -> SimpleExpr {
-        self.bin_oper(BinOper::Like, SimpleExpr::Value(Box::new(v.to_owned())))
+        self.bin_oper(BinOper::Like, SimpleExpr::Value(v.to_owned().into()))
     }
 
     pub fn not_like(self, v: &str) -> SimpleExpr {
-        self.bin_oper(BinOper::NotLike, SimpleExpr::Value(Box::new(v.to_owned())))
+        self.bin_oper(BinOper::NotLike, SimpleExpr::Value(v.to_owned().into()))
     }
 
     /// Express a `IS NULL` expression.
@@ -1071,7 +1071,7 @@ impl Expr {
     /// ```
     pub fn if_null<V>(mut self, v: V) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
     {
         let left = self.left.take();
         Self::func_with_args(
@@ -1132,7 +1132,7 @@ impl Expr {
     #[allow(clippy::wrong_self_convention)]
     pub fn is_in<V, I>(mut self, v: I) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
         I: IntoIterator<Item = V>,
     {
         self.bopr = Some(BinOper::In);
@@ -1194,7 +1194,7 @@ impl Expr {
     #[allow(clippy::wrong_self_convention)]
     pub fn is_not_in<V, I>(mut self, v: I) -> SimpleExpr
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
         I: IntoIterator<Item = V>,
     {
         self.bopr = Some(BinOper::NotIn);
@@ -1734,7 +1734,7 @@ impl SimpleExpr {
         matches!(self, Self::Values(_))
     }
 
-    pub(crate) fn get_values(&self) -> &Vec<Box<dyn QueryValue>> {
+    pub(crate) fn get_values(&self) -> &Vec<Value> {
         match self {
             Self::Values(vec) => vec,
             _ => panic!("not Values"),
