+++ src/query/delete.rs
@@ -3,6 +3,7 @@ use crate::{
     prepare::*,
     query::{condition::*, OrderedStatement},
     types::*,
+    value::*,
     Query, QueryStatementBuilder, SelectExpr, SelectStatement,
 };
 
@@ -186,12 +187,12 @@ impl QueryStatementBuilder for DeleteStatement {
     ///     query.build_collect(MysqlQueryBuilder, &mut collector),
     ///     r#"DELETE FROM `glyph` WHERE `id` = ?"#
     /// );
-    /// assert_eq!(params, vec![Value::Int(Some(1)),]);
+    /// assert_eq!(params, vec![1.into()]);
     /// ```
     fn build_collect<T: QueryBuilder>(
         &self,
         query_builder: T,
-        collector: &mut dyn FnMut(Box<dyn QueryValue>),
+        collector: &mut dyn FnMut(Value),
     ) -> String {
         let mut sql = SqlWriter::new();
         query_builder.prepare_delete_statement(self, &mut sql, collector);
@@ -201,7 +202,7 @@ impl QueryStatementBuilder for DeleteStatement {
     fn build_collect_any(
         &self,
         query_builder: &dyn QueryBuilder,
-        collector: &mut dyn FnMut(Box<dyn QueryValue>),
+        collector: &mut dyn FnMut(Value),
     ) -> String {
         let mut sql = SqlWriter::new();
         query_builder.prepare_delete_statement(self, &mut sql, collector);
