+++ src/query/insert.rs
@@ -1,6 +1,6 @@
 use crate::{
-    backend::QueryBuilder, error::*, prepare::*, types::*, Expr, Query, QueryStatementBuilder,
-    SelectExpr, SelectStatement, SimpleExpr,
+    backend::QueryBuilder, error::*, prepare::*, types::*, value::*, Expr, Query,
+    QueryStatementBuilder, SelectExpr, SelectStatement, SimpleExpr,
 };
 
 /// Insert any new rows into an existing table
@@ -79,10 +79,12 @@ impl InsertStatement {
     /// ```
     /// use sea_query::{tests_cfg::*, *};
     ///
+    /// let values: Vec<Value> = vec![2.1345.into(), "24B".into()];
+    ///
     /// let query = Query::insert()
     ///     .into_table(Glyph::Table)
     ///     .columns(vec![Glyph::Aspect, Glyph::Image])
-    ///     .values(vec![2.1345.into(), "24B".into()])
+    ///     .values(values)
     ///     .unwrap()
     ///     .values_panic(vec![5.15.into(), "12A".into()])
     ///     .to_owned();
@@ -102,7 +104,7 @@ impl InsertStatement {
     /// ```
     pub fn values<V, I>(&mut self, values: I) -> Result<&mut Self>
     where
-        V: Into<Box<dyn QueryValue>>,
+        V: Into<Value>,
         I: IntoIterator<Item = V>,
     {
         let values = values
@@ -167,7 +169,7 @@ impl InsertStatement {
     /// Specify a row of values to be inserted, variation of [`InsertStatement::values`].
     pub fn values_panic<I>(&mut self, values: I) -> &mut Self
     where
-        I: IntoIterator<Item = Box<dyn QueryValue>>,
+        I: IntoIterator<Item = Value>,
     {
         self.values(values).unwrap()
     }
@@ -273,15 +275,15 @@ impl QueryStatementBuilder for InsertStatement {
     /// assert_eq!(
     ///     params,
     ///     vec![
-    ///         Value::Double(Some(3.1415)),
-    ///         Value::String(Some(Box::new(String::from("041080")))),
+    ///         3.1415.into(),
+    ///         "041080".into(),
     ///     ]
     /// );
     /// ```
     fn build_collect<T: QueryBuilder>(
         &self,
         query_builder: T,
-        collector: &mut dyn FnMut(Box<dyn QueryValue>),
+        collector: &mut dyn FnMut(Value),
     ) -> String {
         let mut sql = SqlWriter::new();
         query_builder.prepare_insert_statement(self, &mut sql, collector);
@@ -291,7 +293,7 @@ impl QueryStatementBuilder for InsertStatement {
     fn build_collect_any(
         &self,
         query_builder: &dyn QueryBuilder,
-        collector: &mut dyn FnMut(Box<dyn QueryValue>),
+        collector: &mut dyn FnMut(Value),
     ) -> String {
         let mut sql = SqlWriter::new();
         query_builder.prepare_insert_statement(self, &mut sql, collector);
