+++ src/table/create.rs
@@ -18,7 +18,7 @@ use crate::{
 ///     .col(ColumnDef::new(Char::Character).string().not_null())
 ///     .col(ColumnDef::new(Char::SizeW).integer().not_null())
 ///     .col(ColumnDef::new(Char::SizeH).integer().not_null())
-///     .col(ColumnDef::new(Char::FontId).integer().default(Value::Int(None)))
+///     .col(ColumnDef::new(Char::FontId).integer().default(()))
 ///     .foreign_key(
 ///         ForeignKey::create()
 ///             .name("FK_2e303c3a712662f1fc2a4d0aad6")
