+++ src/types.rs
@@ -1,7 +1,6 @@
 //! Base types used throughout sea-query.
 
-use crate::{expr::*, query::*, CommonSqlQueryBuilder, QueryBuilder};
-use dyn_clonable::*;
+use crate::{expr::*, query::*};
 use std::fmt;
 
 #[cfg(not(feature = "thread-safe"))]
@@ -55,149 +54,6 @@ impl fmt::Debug for dyn Iden {
     fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.unquoted(formatter);
         Ok(())
-    }
-}
-
-/// Indicates that a SQL type is supported for use in queries.
-/// Convert a value to a String for use in queries.
-#[clonable]
-pub trait QueryValue: Clone {
-    /// Returns the value as an escaped string safe for use in queries.
-    fn query_value(&self, query_builder: &dyn QueryBuilder) -> String;
-}
-
-impl std::fmt::Debug for dyn QueryValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.query_value(&CommonSqlQueryBuilder))
-    }
-}
-
-impl std::fmt::Display for dyn QueryValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.query_value(&CommonSqlQueryBuilder))
-    }
-}
-
-impl<T> QueryValue for Option<T>
-where
-    T: QueryValue + Clone,
-{
-    fn query_value(&self, query_builder: &dyn QueryBuilder) -> String {
-        match self {
-            Some(value) => value.query_value(query_builder),
-            None => "NULL".to_string(),
-        }
-    }
-}
-
-macro_rules! into_box_query_value {
-    ($ty: ty) => {
-        impl From<$ty> for Box<dyn QueryValue> {
-            fn from(value: $ty) -> Self {
-                Box::new(value)
-            }
-        }
-
-        impl From<Option<$ty>> for Box<dyn QueryValue> {
-            fn from(value: Option<$ty>) -> Self {
-                Box::new(value)
-            }
-        }
-    };
-}
-
-macro_rules! into_box_query_value_owned {
-    ($ty: ty) => {
-        impl From<$ty> for Box<dyn QueryValue> {
-            fn from(value: $ty) -> Self {
-                Box::new(value.to_owned())
-            }
-        }
-
-        impl From<Option<$ty>> for Box<dyn QueryValue> {
-            fn from(value: Option<$ty>) -> Self {
-                Box::new(value.map(ToOwned::to_owned))
-            }
-        }
-    };
-}
-
-into_box_query_value!(());
-into_box_query_value_owned!(&str);
-into_box_query_value!(String);
-into_box_query_value!(i32);
-into_box_query_value!(i64);
-into_box_query_value!(u32);
-into_box_query_value!(u64);
-into_box_query_value!(f32);
-into_box_query_value!(f64);
-into_box_query_value!(u8);
-into_box_query_value!(Vec<u8>);
-#[cfg(feature = "with-chrono")]
-into_box_query_value!(chrono::DateTime<chrono::FixedOffset>);
-#[cfg(feature = "with-chrono")]
-into_box_query_value!(chrono::NaiveDateTime);
-
-macro_rules! impl_query_value {
-    ($ty: ty) => {
-        impl QueryValue for $ty {
-            fn query_value(&self, _query_builder: &dyn QueryBuilder) -> String {
-                format!("{}", self)
-            }
-        }
-    };
-}
-
-macro_rules! impl_query_value_quoted {
-    ($ty: ty) => {
-        impl QueryValue for $ty {
-            fn query_value(&self, query_builder: &dyn QueryBuilder) -> String {
-                let mut buf = String::new();
-                query_builder.write_string_quoted(self.as_ref(), &mut buf);
-                buf
-            }
-        }
-    };
-}
-
-impl_query_value_quoted!(&str);
-impl_query_value_quoted!(String);
-impl_query_value!(i32);
-impl_query_value!(i64);
-impl_query_value!(u32);
-impl_query_value!(u64);
-impl_query_value!(f32);
-impl_query_value!(f64);
-impl_query_value!(u8);
-
-impl QueryValue for () {
-    fn query_value(&self, _query_builder: &dyn QueryBuilder) -> String {
-        "NULL".to_string()
-    }
-}
-
-impl QueryValue for Vec<u8> {
-    fn query_value(&self, _query_builder: &dyn QueryBuilder) -> String {
-        format!(
-            "x\'{}\'",
-            self.iter()
-                .map(|b| format!("{:02X}", b))
-                .collect::<String>()
-        )
-    }
-}
-
-#[cfg(feature = "with-chrono")]
-impl QueryValue for chrono::DateTime<chrono::FixedOffset> {
-    fn query_value(&self, _query_builder: &dyn QueryBuilder) -> String {
-        format!("\'{}\'", self.format("%Y-%m-%d %H:%M:%S %:z").to_string())
-    }
-}
-
-#[cfg(feature = "with-chrono")]
-impl QueryValue for chrono::NaiveDateTime {
-    fn query_value(&self, _query_builder: &dyn QueryBuilder) -> String {
-        format!("\'{}\'", self.format("%Y-%m-%d %H:%M:%S").to_string())
     }
 }
