+++ tests/postgres/table.rs
@@ -71,11 +71,7 @@ fn create_3() {
             .col(ColumnDef::new(Char::Character).string_len(255).not_null())
             .col(ColumnDef::new(Char::SizeW).integer().not_null())
             .col(ColumnDef::new(Char::SizeH).integer().not_null())
-            .col(
-                ColumnDef::new(Char::FontId)
-                    .integer()
-                    .default(Value::Int(None))
-            )
+            .col(ColumnDef::new(Char::FontId).integer().default(()))
             .foreign_key(
                 ForeignKey::create()
                     .name("FK_2e303c3a712662f1fc2a4d0aad6")
